Refactor visit item to manage programmed status
