+++ package.json
@@ -37,6 +37,7 @@
     "eslint": "^7.32.0",
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "0.72.3",
+    "react-native-clean-project": "^4.0.1",
     "react-native-svg-transformer": "^1.0.0",
     "react-test-renderer": "18.1.0",
     "typescript": "^4.8.3"
