+++ src/components/VisitItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ScrollView, Text, View } from 'react-native';
+import { Text, View } from 'react-native';
 
 import { styles } from '../styles/styles';
 import { Visit } from '../models/visit';
@@ -39,6 +39,12 @@ export const VisitItem = (props: { visit: Visit }) => {
                         <Text style={styles.visitOutcomeText}>NEGATIVO</Text>
                     </View>
                 );
+            case 2:
+                return (
+                    <View style={styles.visitOutcomeButtonNeutral}>
+                        <Text style={styles.visitOutcomeText}>PROGRAMMATA</Text>
+                    </View>
+                );
         }
 
     }
@@ -57,7 +63,7 @@ export const VisitItem = (props: { visit: Visit }) => {
     return (
         <View style={styles.visitItem}>
             <Text style={styles.visitDate}>{(props.visit.date).toLocaleDateString()}</Text>
-            <Text style={{margin: 5, fontStyle: 'italic', color: Colors.black}}>{props.visit.note}</Text>
+            <Text style={{ margin: 5, fontStyle: 'italic', color: Colors.black }}>{props.visit.note}</Text>
             {renderVisitItem(props.visit.status)}
         </View>
     );
